DAO accepts double as target. Wire up ElevationService to use dao.matches, update elevation service tests

+++ src/main/java/com/kyrobot/howfar/data/DataAccessObject.java
@@ -42,7 +42,7 @@ public interface DataAccessObject<T> {
 	 * 	supplied target. Implementers decide what constitutes
 	 * 	a match.
 	 */
-	Stream<T> getMatches(int target, int n);
+	Stream<T> getMatches(double target, int n);
